// src/app/models/request.model.ts
export class RequestModel {
    uid?: string;
    userUid?: string;
    petUid?: string;
    dateBegin?: number;
    viewNotification?: boolean;
    contact?: boolean;
    dateContact?: number;
    dateEnd?: number;
    qts1?: string;
    qts2?: string;
    qts3?: string;
    qts4?: string;
    qts5?: string;
    qts6?: string;
    qts7?: string;
    qts8?: string;
    qts9?: string;
    qts10?: string;
    qts11?: string;
    adopted?: boolean;

}